Fix dark quotes implementation in V2 QuoteView
Since the V2 layout flattens the hierarchy and gets rid of the
intervening background element, we can't do the semi-transparent layers
hack to make the quote itself look different from the message that
quotes it. We merge the two colours of the speaker's chat colour and the
slightly translucent white overlay into one colour and then just set
that as the direct background colour.

Also improve inline documentation for what's going on.